fix sitemap.xml and robots.txt for now v2
