+++ next.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  experimental: {
+    async rewrites() {
+      return [{ source: "/sitemap.xml", destination: "/api/sitemap" }];
+    },
+  },
+};
